move some btc  methods from front helpers to common coin utils
