+++ src/common/utils/coin/btc.ts
@@ -3,7 +3,7 @@ import { BigNumber } from 'bignumber.js'
 import * as bitcoin from 'bitcoinjs-lib'
 import typeforce from 'swap.app/util/typeforce'
 import constants from 'common/helpers/constants'
-import btcHelper from '../../../front/shared/helpers/btc'
+import request from 'common/utils/request'
 
 // Use front API config
 import { default as TESTNET } from '../../../front/config/testnet/api'
@@ -756,26 +756,239 @@ const estimateFeeRateEARNCOM = (options) => {
 }
 
 const estimateFeeRate = async (options) => {
+  const { speed = 'fast', method } = options
+
+  if (method === 'swap'){
+    try {
+      return await estimateFeeRateBLOCKCYPHER(options)
+    } catch (err) {
+      console.error(`EstimateFeeError: BLOCKCYPHER_API ${err.message}, trying EARN.COM...`)
+      return await estimateFeeRateEARNCOM(options)
+    }
+  }
+  const link = 'https://wiki.swaponline.io/blockcyper.php'
+  const defaultRate = {
+    slow: 5 * 1e3,
+    normal: 15 * 1e3,
+    fast: 30 * 1e3,
+  }
+
+  let apiResult
+
   try {
-    return await estimateFeeRateBLOCKCYPHER(options)
+    // api returns sotoshi in 1 kb
+    apiResult = await request.get(link, { cacheResponse: 60000 })
   } catch (err) {
-    console.error(`EstimateFeeError: BLOCKCYPHER_API ${err.message}, trying EARN.COM...`)
-    return await estimateFeeRateEARNCOM(options)
+    console.error("Can't fetch ${link}")
+    return defaultRate[speed]
+  }
+
+  const apiSpeeds = {
+    slow: 'low_fee_per_kb',
+    normal: 'medium_fee_per_kb',
+    fast: 'high_fee_per_kb',
+  }
+
+  const apiSpeed = apiSpeeds[speed] || apiSpeeds.normal
+  const apiRate = new BigNumber(apiResult[apiSpeed])
+
+  return apiRate.isGreaterThanOrEqualTo(constants.TRANSACTION.DUST_SAT)
+    ? apiRate.toNumber()
+    : defaultRate[speed]
+}
+
+enum Network {
+  mainnet = "mainnet",
+  testnet = "testnet",
+}
+
+enum AddressType {
+  p2pkh = "P2PKH",
+  p2sh = "P2SH",
+  p2wpkh = "P2WPKH",
+  p2wsh = "P2WSH",
+}
+
+const addressTypes: { [key: number]: { type: AddressType, network: Network } } = {
+  0x00: {
+    type: AddressType.p2pkh,
+    network: Network.mainnet,
+  },
+
+  0x6f: {
+    type: AddressType.p2pkh,
+    network: Network.testnet,
+  },
+
+  0x05: {
+    type: AddressType.p2sh,
+    network: Network.mainnet,
+  },
+
+  0xc4: {
+    type: AddressType.p2sh,
+    network: Network.testnet,
+  },
+};
+
+const getAddressType = (address: string) => {
+  const prefix = address.substr(0, 2);
+  let version;
+  let data;
+  let addressType;
+
+  if (prefix === 'bc' || prefix === 'tb') {
+    const { data: benchVersion } = bitcoin.address.fromBech32(address)
+    data = benchVersion;
+    return addressType = data.length === 20 ? AddressType.p2wpkh : AddressType.p2wsh;
+
+  } else {
+    const { version: baseVersion } = bitcoin.address.fromBase58Check(address)
+    version = baseVersion
+    let { type } = addressTypes[version]
+    if (!type) {
+      type = AddressType.p2pkh;
+      console.warn(`Unknown version '${version}' for address '${address}'.`)
+    }
+    return addressType = type
   }
 }
 
-const calculateTxSize = async (params) => {
+// getByteCount({'MULTISIG-P2SH:2-4':45},{'P2PKH':1}) Means "45 inputs of P2SH Multisig and 1 output of P2PKH"
+// getByteCount({'P2PKH':1,'MULTISIG-P2SH:2-3':2},{'P2PKH':2}) means "1 P2PKH input and 2 Multisig P2SH (2 of 3) inputs along with 2 P2PKH outputs"
+const getByteCount = (inputs, outputs) => {
   const { TRANSACTION } = constants
-  const { address, NETWORK } = params
+  let totalWeight = 0
+  let hasWitness = false
+  let inputCount = 0
+  let outputCount = 0
+  // assumes compressed pubkeys in all cases.
+  const types = {
+    'inputs': {
+      'MULTISIG-P2SH': TRANSACTION.MULTISIG_P2SH_IN_SIZE * 4,
+      'MULTISIG-P2WSH': TRANSACTION.MULTISIG_P2WSH_IN_SIZE + (41 * 4),
+      'MULTISIG-P2SH-P2WSH': TRANSACTION.MULTISIG_P2SH_P2WSH_IN_SIZE + (76 * 4),
+      'P2PKH': TRANSACTION.P2PKH_IN_SIZE * 4,
+      'P2WPKH': TRANSACTION.P2WPKH_IN_SIZE + (41 * 4),
+      'P2SH-P2WPKH': TRANSACTION.P2SH_P2WPKH_IN_SIZE + (64 * 4),
+    },
+    'outputs': {
+      'P2SH': TRANSACTION.P2SH_OUT_SIZE * 4,
+      'P2PKH': TRANSACTION.P2PKH_OUT_SIZE * 4,
+      'P2WPKH': TRANSACTION.P2WPKH_OUT_SIZE * 4,
+      'P2WSH': TRANSACTION.P2WSH_OUT_SIZE * 4,
+    },
+  }
 
-  const unspents = await fetchUnspents({
+  const checkUInt53 = (n) => {
+    if (n < 0 || n > Number.MAX_SAFE_INTEGER || n % 1 !== 0) throw new RangeError('value out of range')
+  }
+
+  const varIntLength = (number) => {
+    checkUInt53(number)
+
+    return (
+      number < 0xfd ? 1
+        : number <= 0xffff ? 3
+          : number <= 0xffffffff ? 5
+            : 9
+    )
+  }
+
+  Object.keys(inputs).forEach((key) => {
+    checkUInt53(inputs[key])
+    if (key.slice(0, 8) === 'MULTISIG') {
+      // ex. "MULTISIG-P2SH:2-3" would mean 2 of 3 P2SH MULTISIG
+      const keyParts = key.split(':')
+      if (keyParts.length !== 2) throw new Error(`invalid input: ${key}`)
+      const newKey = keyParts[0]
+      const mAndN = keyParts[1].split('-').map((item) => parseInt(item))
+
+      totalWeight += types.inputs[newKey] * inputs[key]
+      const multiplyer = (newKey === 'MULTISIG-P2SH') ? 4 : 1
+      totalWeight += ((73 * mAndN[0]) + (34 * mAndN[1])) * multiplyer * inputs[key]
+    } else {
+      totalWeight += types.inputs[key] * inputs[key]
+    }
+    inputCount += inputs[key]
+    if (key.indexOf('W') >= 0) hasWitness = true
+  })
+
+  Object.keys(outputs).forEach((key) => {
+    checkUInt53(outputs[key])
+    totalWeight += types.outputs[key] * outputs[key]
+    outputCount += outputs[key]
+  })
+
+  if (hasWitness) totalWeight += 2
+
+  totalWeight += 8 * 4
+  totalWeight += varIntLength(inputCount) * 4
+  totalWeight += varIntLength(outputCount) * 4
+
+  return Math.ceil(totalWeight / 4)
+}
+
+type CalculateTxSizeParams = {
+  txIn?: number
+  txOut?: number
+  address?: string,
+  NETWORK?: string,
+  method?: string
+  fixed?: boolean
+  toAddress?: string
+  serviceFee?: {
+    address: any,
+    min: any,
+    fee: any
+  }
+}
+
+const calculateTxSize = async (params: CalculateTxSizeParams) => {
+  const { TRANSACTION } = constants
+  const {
     address,
     NETWORK,
-  })
-  const txOut = 2
-  const txIn = unspents.length
-  let txSize = 226 // default tx size for 1 txIn and 2 txOut
+    method = 'send',
+    txIn,
+    txOut,
+    fixed,
+    toAddress,
+    serviceFee,
+  } = params
 
+  if (method === 'swap') {
+    const unspents = await fetchUnspents({
+      address,
+      NETWORK,
+    })
+    const txOut = 2
+    const txIn = unspents.length
+    let txSize = 400 // default tx size for swap method
+
+    if (txIn > 0) {
+      txSize =
+        txIn * TRANSACTION.P2PKH_IN_SIZE +
+        txOut * TRANSACTION.P2PKH_OUT_SIZE +
+        (TRANSACTION.TX_SIZE + txIn - txOut)
+    }
+
+    console.group('Common > coin >%c btc > calculateTxSize (swap)', 'color: green;')
+    console.log('params: ', params)
+    console.log('tx size: ', txSize)
+    console.groupEnd()
+
+    return txSize
+  }
+
+  const defaultTxSize = 226 // default tx size for 1 txIn and 2 txOut
+  if (fixed) {
+    return defaultTxSize
+  }
+
+  let txSize = defaultTxSize
+  // general formula
+  // (<one input size> × <number of inputs>) + (<one output size> × <number of outputs>) + <tx size>
   if (txIn > 0) {
     txSize =
       txIn * TRANSACTION.P2PKH_IN_SIZE +
@@ -783,50 +996,167 @@ const calculateTxSize = async (params) => {
       (TRANSACTION.TX_SIZE + txIn - txOut)
   }
 
+  if (method === 'send_multisig') {
+    let outputs = {
+      'P2SH': 1,
+    }
+    const toAddressType = toAddress ? getAddressType(toAddress) : "P2PKH";
+    outputs[toAddressType] = ++outputs[toAddressType] || 1;
+
+    if (serviceFee) {
+      const adminAddressType = getAddressType(serviceFee.address);
+      outputs[adminAddressType] = ++outputs[adminAddressType] || 1;
+    }
+    txSize = getByteCount(
+      { 'MULTISIG-P2SH:2-2': 1 },
+      outputs
+    )
+  }
+
+  if (method === 'send_2fa') {
+    let outputs = {
+      'P2SH': 1,
+    }
+    const toAddressType = toAddress ? getAddressType(toAddress) : "P2PKH";
+    outputs[toAddressType] = ++outputs[toAddressType] || 1;
+
+    if (serviceFee) {
+      const adminAddressType = getAddressType(serviceFee.address);
+      outputs[adminAddressType] = ++outputs[adminAddressType] || 1;
+    }
+    txSize = getByteCount(
+      { 'MULTISIG-P2SH:2-3': txIn },
+      outputs
+    )
+  }
+
   console.group('Common > coin >%c btc > calculateTxSize', 'color: green;')
   console.log('params: ', params)
-  console.log('tx size: ', txSize)
+  console.log('txSize: ', txSize)
   console.groupEnd()
 
   return txSize
 }
 
-const estimateFeeValue = async (options) => {
+type EstimateFeeValueParams = {
+  method?: string
+  speed: 'fast' | 'normal' | 'slow'
+  feeRate?: number
+  inSatoshis?: boolean
+  address: string
+  txSize?: number
+  fixed?: boolean
+  amount?: number
+  toAddress?: string
+  swapUTXOMethod?: 'withdraw' | 'deposit'
+  moreInfo?: boolean
+  serviceFee?: {
+    address: any,
+    min: any,
+    fee: any
+  }
+  NETWORK?: 'MAINNET' | 'TESTNET'
+}
+
+const estimateFeeValue = async (params: EstimateFeeValueParams) => {
   const {
     feeRate: _feeRate,
     inSatoshis,
     speed,
     address,
     txSize: _txSize,
+    fixed,
+    method,
+    amount,
+    toAddress,
+    swapUTXOMethod,
+    moreInfo,
+    serviceFee,
     NETWORK,
-  } = options
+  } = params
 
-  let calculatedFeeValue
+  if(method === 'swap') {
+    let calculatedFeeValue
 
-  if (!_txSize && !address) {
-    calculatedFeeValue = new BigNumber(constants.TRANSACTION.DUST_SAT).multipliedBy(1e-8)
-  } else {
-    const txSize = _txSize || await calculateTxSize({ address, NETWORK })
-    const feeRate = _feeRate || await estimateFeeRate({ speed, NETWORK })
+    if (!_txSize && !address) {
+      calculatedFeeValue = new BigNumber(constants.TRANSACTION.DUST_SAT).multipliedBy(1e-8)
+    } else {
+      const txSize = _txSize || await calculateTxSize({ address, method, NETWORK })
+      const feeRate = _feeRate || await estimateFeeRate({ speed, NETWORK })
 
-    calculatedFeeValue = BigNumber.maximum(
-      constants.TRANSACTION.DUST_SAT,
-      new BigNumber(feeRate)
-        .multipliedBy(txSize)
-        .div(1024)
-        .dp(0, BigNumber.ROUND_HALF_EVEN),
-    )
+      calculatedFeeValue = BigNumber.maximum(
+        constants.TRANSACTION.DUST_SAT,
+        new BigNumber(feeRate)
+          .multipliedBy(txSize)
+          .div(1024)
+          .dp(0, BigNumber.ROUND_HALF_EVEN),
+      )
+    }
+
+    const finalFeeValue = inSatoshis
+      ? calculatedFeeValue.toString()
+      : calculatedFeeValue.multipliedBy(1e-8).toString()
+
+    console.group('Common > coin >%c btc > estimateFeeValue', 'color: green;')
+    console.log('fee value: ', finalFeeValue)
+    console.groupEnd()
+
+    return finalFeeValue
+  }
+  let unspents: IBtcUnspent[] = await fetchUnspents({
+    address,
+    NETWORK,
+  })
+  // if user have some amount then try to find "better" UTXO for this
+  if (amount) {
+    unspents = await prepareUnspents({ amount, unspents })
   }
 
-  const finalFeeValue = inSatoshis
-    ? calculatedFeeValue.toString()
-    : calculatedFeeValue.multipliedBy(1e-8).toString()
+  // one input for output from the script when swapping
+  const txIn = swapUTXOMethod === 'withdraw' ? 1 : unspents.length
+  // 2 = recipient input + sender input (for a residue)
+  // 3 = the same inputs like higher + input for admin fee
+  const txOut = serviceFee
+    ? method === 'send'
+      ? 3
+      : 2
+    : 2
 
-  console.group('Common > coin >%c btc > estimateFeeValue', 'color: green;')
-  console.log('fee value: ', finalFeeValue)
-  console.groupEnd()
+  const feeRate = _feeRate || await estimateFeeRate({ speed })
+  const txSize = _txSize || await calculateTxSize({
+    fixed,
+    method,
+    txIn,
+    txOut,
+    toAddress
+  })
+
+  const calculatedFeeValue = BigNumber.maximum(
+    constants.TRANSACTION.DUST_SAT,
+    new BigNumber(feeRate)
+      .multipliedBy(txSize)
+      .div(1024) // divide by one kilobyte
+      .dp(0, BigNumber.ROUND_HALF_EVEN),
+  )
+
+  const SATOSHI_TO_BITCOIN_RATIO = 0.000_000_01
+
+  const finalFeeValue = inSatoshis
+    ? calculatedFeeValue.toNumber()
+    : calculatedFeeValue.multipliedBy(SATOSHI_TO_BITCOIN_RATIO).toNumber()
+
+  if (moreInfo) {
+    return {
+      fee: calculatedFeeValue.multipliedBy(SATOSHI_TO_BITCOIN_RATIO).toNumber(),
+      satoshis: calculatedFeeValue.toNumber(),
+      txSize,
+      feeRate,
+      unspents,
+    }
+  }
 
   return finalFeeValue
+
 }
 
 const prepareFees = async ({
@@ -859,7 +1189,7 @@ const prepareFees = async ({
   try {
     feeValue = feeValue ?
       new BigNumber(feeValue).multipliedBy(1e8).toNumber() :
-      await btcHelper.estimateFeeValue({ // after refactoring helpers need use here method
+      await estimateFeeValue({
         inSatoshis: true,
         speed,
         method,
@@ -1004,6 +1334,8 @@ export default {
   getTransactionBlocyper,
 
   estimateFeeValue,
+  estimateFeeRate,
+  calculateTxSize,
   getCore,
 
   prepareUnspents,
